Fix client certificate chain with aktualizr
