+++ setup.cfg
@@ -3,7 +3,6 @@ follow_imports=silent
 strict_optional=True
 show_error_context=True
 ignore_missing_imports=True
-disallow_incomplete_defs=True
 disallow_untyped_calls=True
 disallow_untyped_defs=True
 warn_return_any=True
