+++ src/datatypes/metadata_test.py
@@ -1,13 +1,21 @@
 import json
 import os
+import pytest
+
+from typing import Any, Callable, Type
 
 from .metadata import Metadata
-from ..errors import InvalidKeyIdError, MissingFieldError, UnknownRoleError
-from ..utils import assert_raises, canonical
+from ..errors import Error, InvalidKeyIdError, MissingFieldError, UnknownRoleError
+from ..utils import canonical
 
 
 META_DIR = "fixtures/metadata"
 KEY_ID = "0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef"
+
+def assert_raises(e: Type[Error], f: Callable, *args: Any) -> None:
+    """Capture an expected error while testing."""
+    with pytest.raises(e):
+        f(*args)
 
 
 def test_read_metadata() -> None:
